refactor: make pycleo compatible with new configuration library
